Restore qrexec policy from .rpmsave files
When packages are upgrade to a version that no longer include legacy
policy, RPM may decide to move the (user-modified) policy to a separate
file with .rpmsave extension. Such policy file will be ignored, and thus
breaks user's qrexec policy. If that happen to a qubes.InputKeyboard
service, it may render system hard to use.

While removing legacy policy is desired, it should be done with
migration tool that is separate. Stage 3 of release upgrade should not
change policy semantics, so restore those files.

QubesOS/qubes-issues#7832